SmartSkip.lua
fix: autoskip not showing consecutive chapters that will be auto-skipped (when playlist-next) is detected.

With all the added features inside autoskip, the function  "chapterskip" has a lot of repeated code, it needs a clean-up to utilize functions instead of repeating the same code. 

My priority now, however is to get it working despite lesser code quality.